Middleware verificando token done

+++ controllers/books.controller.js
@@ -21,7 +21,7 @@ const BooksController = {
   // Crear un libro
   createBook: (req, res) => {
     if (!(req.body.title && req.body.author && req.body.pageNumber)) {
-      res.status(400).send({ Error: ' El usuario no se agrego! , Intentalo de nuevo ( Revisa los nombres )' });
+      res.status(400).send({ Error: ' El LIBRO no se agrego!! , Intentalo de nuevo ( Revisa los nombres )' });
     } else {
       const title = req.body.title;
       const author = req.body.author;
@@ -38,20 +38,19 @@ const BooksController = {
         .catch(err => {
           res.status(500).send({ message: "Trono la Base de Datos o la creación del libro! D:" });
         })
-
     }
   },
 
   // Encontrar y leer un libro
   readOneBook: (req, res) => {
-    let id = req.body.id;
+    let id = req.param.id;
     BookModel
       .findOne({ _id: id })
       .then(data => {
         if (data) {
           res.status(200).send({ data: data }).end();
         } else {
-          res.status(400).send({ Error: 'No pudimos encontrar el libro, intenta otro ID' });
+          res.status(404).send({ Error: 'No pudimos encontrar el libro, intenta otro ID' });
         }
       })
       .catch(err => {
@@ -62,13 +61,13 @@ const BooksController = {
 
   // Sobre escribir un libro
   updateBook: (req, res) => {
-    if (!(req.body.title && req.body.author && req.body.pageNumber && req.body.id)) {
-      res.status(400).send({ Error: 'Ingresa todos los parametros!' });
+    if (!(req.body.title && req.body.author && req.body.pageNumber && req.params.id)) {
+      res.status(400).send({ Error: 'Ingresa todos los parametros! Asegurate de tener title,author pageNumber y ID!!!' });
     } else {
       const title = req.body.title;
       const author = req.body.author;
       const pageNumber = req.body.pageNumber;
-      const id = req.body.id;
+      const id = req.params.id;
       BookModel
         .findOneAndUpdate({
           _id: id
@@ -78,7 +77,7 @@ const BooksController = {
             pageNumber
           })
         .then(data => {
-          res.status(201).send({ data: data }).end(); //regresa los datos del libro que se esta sobreescribiendo
+          res.status(200).send({ data: data }).end(); //regresa los datos del libro que se esta sobreescribiendo
         })
         .catch(err => {
           res.status(500).send({ message: "Trono la Base de Datos! D:" });
@@ -88,16 +87,20 @@ const BooksController = {
 
   // Borrar un libro con ID
   deleteBook: (req, res) => {
-    let id = req.body.id;
+    if(req.params.id){
+    let id = req.params.id;
     BookModel
       .deleteOne({ _id: id })
       .then(data => {
-        res.status(200).send({ message: "Libro borrado!" }); //manda mensaje de borrado y el libro que se borro
+        res.status(200).send({ message: "Libro borrado 4eva & eva!" }); //manda mensaje de borrado y el libro que se borro
       })
       .catch(err => {
         res.status(500).send({ message: "Trono la Base de Datos! D:" });
       });
+    }else{
+      res.status(400).send({message: "¡Oye! Revisa si mandas ID "})
+    }
   }
 
-}
+};
 module.exports = BooksController;
